feat: Add HomeScreen component with property list
This commit introduces a new HomeScreen component to the project, which
serves as the main landing page. The HomeScreen includes a property list
that displays a collection of properties.

By adding the HomeScreen and property list, users can easily browse and
explore available properties right from the start of the application.